requirements.txt
streamlit
opencv-python-headless
google-cloud-vision
numpy
Pillow
